Corrected errors and mistakes in tests, ammended file paths. Issues persist with running tests locally but they are all referring to erroneous file path (eilnn.IMAGES_DIR)

+++ eilnn/utils.py
@@ -43,7 +43,7 @@ def load_grayscale(grayscale_path):
         ]
     )
     
-    middle_slice = image_stack.shape[0]/2
+    middle_slice = int(image_stack.shape[0]/2)
     plt.imshow(image_stack[middle_slice])
     print('Stack Dimensions ', str(image_stack.shape))
